Code improvements with additional compiler flags
 - Code improvements to ensure no compiler warnings and errors.
 - Removal of unused variables.

+++ test_pool/mpam/operating_system/test_mpam002.c
@@ -28,7 +28,7 @@
 static void payload(void)
 {
     uint32_t llc_index;
-    uint32_t cache_identifier;
+    uint64_t cache_identifier;
     uint32_t msc_node_cnt;
     uint32_t rsrc_node_cnt;
     uint32_t msc_index, rsrc_index;
